Introduce db, dto, msg module
