+++ cs2posts/bot/constants.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 REQUESTS_TIMEOUT = 3
-TELEGRAM_MAX_MESSAGE_LENGTH = 4096
 
 WELCOME_MESSAGE_ENGLISH = """\
 <b>Welcome to the Counter-Strike 2 Post Bot!</b>\n
