--- cs2posts/bot/decorators.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-import logging
-
-logger = logging.getLogger(__name__)
-
-# TODO: Add decorators here.
