+++ cs2posts/db/__init__.py
@@ -1,0 +1,11 @@
+from __future__ import annotations
+
+from .db_chats import ChatDatabase
+from .db_posts import PostDatabase
+from .db_sqlite import SQLite
+
+__all__ = [
+    'PostDatabase',
+    'ChatDatabase',
+    'SQLite',
+]
