+++ cs2posts/dto/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from .chats import Chat
+from .post import Post
+
+__all__ = [
+    "Chat",
+    "Post",
+]
