+++ cs2posts/dto/post.py
