+++ cs2posts/msg/constants.py
@@ -1,0 +1,2 @@
+from __future__ import annotations
+TELEGRAM_MAX_MESSAGE_LENGTH = 4096
