+++ save.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 from cs2posts.crawler import CounterStrike2Crawler
 from cs2posts.cs2 import CounterStrike2Posts
-from cs2posts.post import Post
+from cs2posts.dto.post import Post
 
 
 class PostNotFound(Exception):
